1)fixed the data type in location table

+++ SQL/init_db.sql
@@ -71,8 +71,8 @@ CREATE TABLE REPORT (
 -- table Location
 CREATE TABLE LOCATION (
     Incident_ID INT             NOT NULL,
-	Longitude INT               NOT NULL,
-	Latitude  INT               NOT NULL,
+	Longitude REAL              NOT NULL,
+	Latitude  REAL              NOT NULL,
 	Intersection VARCHAR(255)   NOT NULL,
     FOREIGN KEY (Incident_ID) REFERENCES INCIDENTS(Incident_ID)
 );
